Re-organize billing package
