+++ src/main/java/com/treasuredata/android/TreasureData.java
@@ -8,7 +8,7 @@ import android.content.pm.PackageInfo;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
-import com.treasuredata.android.internal.PurchaseEventActivityLifecycleTracker;
+import com.treasuredata.android.billing.internal.PurchaseEventActivityLifecycleTracker;
 import io.keen.client.java.KeenClient;
 import org.komamitsu.android.util.Log;
